internals refactor
