+++ config/test-setup.ts
@@ -1,0 +1,7 @@
+import {before} from 'mocha'
+
+before(async () => {
+  if (typeof globalThis.fetch === 'undefined') {
+    await import('../src/polyfill')
+  }
+})
